requirements.txt
langchain==1.1.0
langchain-community
langchain-classic
langchain-google-genai
google-generativeai
langchain-chroma
langchain-huggingface
transformers
huggingface-hub
python-dotenv
streamlit
fastapi

numpy
scikit-learn
unstructured==0.18.21
unstructured-client==0.42.4
nstructured_inference==1.1.2
unstructured[all-docs]